ADD: weight in dat fields are in tons but kg can bei give like 1.350
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@5777 8aca7d54-2c30-db11-9de9-000461428c89

+++ besch/vehikel_besch.h
@@ -245,7 +245,7 @@ public:
 	uint16 get_loading_time() const { return loading_time; } // ms per full loading/unloading
 	uint32 get_preis() const { return preis; }
 	sint32 get_geschw() const { return geschw; }
-	uint16 get_gewicht() const { return gewicht; }
+	uint32 get_gewicht() const { return gewicht; }
 	uint16 get_axle_load() const { return axle_load; }
 	uint32 get_leistung() const { return leistung; }
 	uint16 get_betriebskosten() const { return running_cost; }
